requirements.txt
# Requirements for Adversarial Noise Project
# This file contains all the necessary Python packages to run the adversarial attack examples

# Core deep learning framework
torch>=2.0.0
torchvision>=0.15.0

# Numerical computing and data manipulation
numpy>=1.21.0
scipy>=1.7.0

# Image processing
Pillow>=8.3.0
opencv-python>=4.5.0

# Visualization and plotting
matplotlib>=3.5.0
seaborn>=0.11.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

# Additional utilities
tqdm>=4.62.0
requests>=2.25.0

# Optional: GPU support (if using CUDA)
# torch-cuda>=2.0.0  # Uncomment if you need CUDA support

# Development and testing (optional)
pytest>=6.0.0
black>=21.0.0
flake8>=3.9.0 